requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-dotenv
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
streamlit
langchain
langchain-openai
langchain-huggingface
langchain-llms
langchain-community
chromadb
pypdf
pdf2image
pytesseract
Pillow
python-multipart
sentence-transformers
transformers
torch